backend/lambdas/problem-generator/requirements.txt
google-generativeai>=0.8.4
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0
gradio>=4.0.0